+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow attachments for helper tasks
 - Bulk clone for helper tasks
 - Mass publish for unpublished helper tasks
+- Show task title in browser tab for helper task detail page
 
 ### Changed